refactor: extract task_index_label, split tests into co-located files
- Extract addContestNameToTaskIndex + isAojContest to task_index_label.ts
- Split contest.test.ts into co-located test files:
  - priority.test.ts (contest priority tests)
  - labels/index.test.ts (contest name label tests)
  - labels/universities.test.ts (university label tests)
  - task_index_label.test.ts (task index label tests)
- contest.ts is now a pure re-export module
