+++ src/lib/contests/utils/contest.ts
@@ -1,6 +1,3 @@
-import { regexForJag, regexForAojUniversity, aojCoursePrefixes } from './prefixes';
-import { getContestNameLabel } from './labels/index';
-
 export {
   regexForJag,
   regexForAojUniversity,
@@ -15,23 +12,4 @@ export { getPastContestLabel, PAST_TRANSLATIONS } from './labels/past';
 export { getJoiContestLabel } from './labels/joi';
 export { getAtCoderUniversityContestLabel } from './labels/universities';
 export { getAojContestLabel } from './labels/aoj';
-
-export const addContestNameToTaskIndex = (contestId: string, taskTableIndex: string): string => {
-  const contestName = getContestNameLabel(contestId);
-
-  if (isAojContest(contestId)) {
-    return `AOJ ${taskTableIndex}${contestName}`;
-  }
-
-  return `${contestName} - ${taskTableIndex}`;
-};
-
-function isAojContest(contestId: string): boolean {
-  return (
-    aojCoursePrefixes.has(contestId) ||
-    contestId.startsWith('PCK') ||
-    regexForJag.test(contestId) ||
-    contestId.startsWith('ICPC') ||
-    regexForAojUniversity.test(contestId)
-  );
-}
+export { addContestNameToTaskIndex } from './task_index_label';
